lib: fixed pas_malloc() wrapper definition - should use PtrUInt / size_t everywhere for consistency

+++ res/static/liblizard/lib/fse.h
@@ -46,10 +46,10 @@ extern "C" {
 
 #if !defined(pas_malloc)
 //AB
-void *pas_malloc(int size);
+void *pas_malloc(size_t size);
 void pas_free(void *ptr);
-void *pas_realloc(void *ptr, int size);
-int pas_malloc_usable_size(void *ptr); 
+void *pas_realloc(void *ptr, size_t size);
+size_t pas_malloc_usable_size(void *ptr); 
 #endif
 
 /*-*****************************************
